refactored screens package
