+++ navigation/src/main/java/com/smileidentity/navigation/ui/OrchestratedCaptureScreen.kt
@@ -8,7 +8,7 @@ import com.ramcosta.composedestinations.generated.navigation.destinations.Orches
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
 import com.smileidentity.ml.states.IdentityScanState
-import com.smileidentity.ui.SmileIDCaptureScreen
+import com.smileidentity.ui.screens.SmileIDCaptureScreen
 import com.smileidentity.ui.previews.DevicePreviews
 import com.smileidentity.ui.previews.ThemePreviews
