+++ navigation/src/main/java/com/smileidentity/navigation/ui/OrchestratedPreviewScreen.kt
@@ -7,7 +7,7 @@ import com.ramcosta.composedestinations.annotation.RootGraph
 import com.ramcosta.composedestinations.generated.navigation.destinations.OrchestratedProcessingScreenDestination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
-import com.smileidentity.ui.SmileIDPreviewScreen
+import com.smileidentity.ui.screens.SmileIDPreviewScreen
 import com.smileidentity.ui.previews.DevicePreviews
 import com.smileidentity.ui.previews.ThemePreviews
