+++ navigation/src/main/java/com/smileidentity/navigation/ui/OrchestratedProcessingScreen.kt
@@ -7,7 +7,7 @@ import com.ramcosta.composedestinations.annotation.RootGraph
 import com.ramcosta.composedestinations.generated.navigation.destinations.OrchestratedCaptureScreenDestination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
-import com.smileidentity.ui.SmileIDProcessingScreen
+import com.smileidentity.ui.screens.SmileIDProcessingScreen
 import com.smileidentity.ui.previews.DevicePreviews
 import com.smileidentity.ui.previews.ThemePreviews
