+++ ui/src/androidTest/java/com/smileidentity/ui/SmileIDCaptureScreenTest.kt
@@ -6,6 +6,7 @@ import androidx.compose.ui.test.assertHasClickAction
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import com.smileidentity.ml.states.IdentityScanState
+import com.smileidentity.ui.screens.SmileIDCaptureScreen
 import org.junit.Rule
 import org.junit.Test
